Force kill OpenVPN process on stopping service
This is necessary because the OpenVPN process does not seem to actually exit until the service action queue is empty.  On a service restart, this will mean that the process is still running when we try to start it back up - so it will bail out.  However, since it has already done all of its cleaning up, we'll just force kill it before considering the service stopped.