miniupnpd: check rule desc when removing expired rules
This is importaint since we keep expiration time as part of the description.
When the rule is renewed by the client, we update the rule desc with new
expiration time, but the nextnatpmptoclean_timestamp variable stays unchanged.
When we reach time stored in nextnatpmptoclean_timestamp, CleanExpiredNATPMP()
is called, but without this check the rule for the same proto/port will be
deleted even if it's not yet expired.