Busybox: fix "bus error" in e2fsck (wrong but working)
The resolve_mount_spec() is not correctly used in other applets too,
but hey - mount/swaponoff/findfs don't care, and don't even try to
free allocated strings. So let's wait until the original issue is
(and if) corrected upstream, and then add a proper fix for e2fsck.

For now this works since we don't enable FEATURE_CLEAN_UP.