USB support: make after-mount and before-unmount scripts to work the same way in 2.4 and 2.6-based builds
In kernel 2.6-based builds the after-mount script was called for every partition, while in 2.4 it was
only called once per disk. Before-unmount script was only executed once per disk, and only if the drive
was still plugged in.
Change it so in both targets all after-mount and before-unmount scripts are called for every partition.
The scripts now accept one parameter - the mountpoint. Fix coding style in a few places while making
this change.