rc: rework signals processing, fix for missed signals
The problems with signals in init.  pause() was used in a manner
that is specifically warned against, which can lead to lost signals
and deadlocks.  The bug is that the code does a test and then, based
on that test, suspends.  There is a timing window between the test and
the pause().   The suggested revision was to replace pause() with
sigsuspend().  That doesn't work in Tomato, though.  Because almost
all of the time, it doesn't suspend in the pause() - it suspends in
the select() in run_shell.  The other problem was that the signal
handlers stored the signal into a single variable.  If multiple
signals come at the same time, the earlier one(s) will be over-written
by the last one, and will be lost.

All-in-all, there can be "interesting" failures, depending on random
timing.

This patch fixes that. There were a couple of ways to fix it.  I
chose the one with the simplest logic.

The one thing I cannot test is the "run_shell" feature, since I do not
have a console TTY connection on my router. However, I did test it on
a desktop PC and it seems to work properly.