rc: allow syslog (service logging) to read options from /etc/syslogd.cfg
Note that is's not ".conf" in order to avoid confusion with the REAL syslogd.
Options are: rotate_size  num_of_rotations_to_keep   output_filename

All on the first line of the file, separated by spaces. A "-" means to use the
Tomato default for that number.

You have to create that file,
for example:	echo "500 10 /mnt/TOSH_120/z.logs/syslog" >/etc/syslogd.cfg
then execute:	service logging restart

It won't use the new logfile unless it can write to it in start_syslog.
So the restart command needs to be in a script that is executed after the disk is mounted.