httpd: GUI enhancement to display the actual logfile, named in /etc/syslogd.cfg
Commit 2823153e7b8de1a9b0c481bb9a69b23ac54c9892 allowed syslog to read options
from /etc/syslogd.cfg and to specify alternative log file location.

This completes the change by modifying the GUI to display the actual logfile,
named in /etc/syslogd.cfg. It will correctly work with old rotated backup
logfiles, no limit on the number of them.

Added the arbitrary limit on the number of log lines to display (up to 2000)
to the original patch - to prevent loading potentially a huge amount of data
into the browser.