iptables: fix multiport match for kernel 2.4
Ref: https://dev.openwrt.org/changeset/19761
Since the netfilter framework in Linux 2.4 lacks support for target and match revisioning,
iptables falls back to the v0 implementation of multiport while only v1 is implemented on
the kernel side, leading to an invalid argument error when trying to insert rules using
multiport. This patch removes v0 support from the userspace match library to enforce the
usage of the v1 match implementation.