Web UI: added new page to display per-connection transfer rates
This is done by using the bytes= counts in conntrack. The table is dumped to a tmpfile,
the function waits x seconds, then re-reads the table making a effort to match current
to previous entries in the table, and from that estimates the data transfer rate for each
connection. Only connections with rates above a certain threshold are displayed.