nvram [K24 part]: auto-detect the nvram size
This is first patch from the set for phase 1 of expanding the
NVRAM size to be more than 32KB.

This is phase 1, which is the prep phase.  This is all the code
**except** for actually increasing the NVRAM size. Both application
level code & kernel code.

It will probe to auto-detect the size of the nvram area, so it will
work for both 32KB and 60KB nvram size.

It accurately reports the nvram size and un-used space, nothing about
this is hardcoded.

This also partially reverts commit 589975b14327b626d4eb8aae3fa6a25ce6464c1d