BCM SDK 5.10.147.x: rework WAR for oops'es in et driver
Broadcom ET and CTF drivers are using mac_len member of sk_buff structure
for internal purposes, causing kernel oops'es when something else in the
kernel populates this field.

This commit replaces previous WAR 2fbfdd856fdd50152fc742925a4e5ee2b5fc56bf
which made ET to use different (unused) skbuff member, but since CTF is
a closed source driver we can't change it, and when loaded it was causing
the same problems.

Unfortunately, we have to deal with what Broadcom provides, so better to
change something we have sources for, even if these changes are more
intrusive...