nvram [K26]: try to avoid a concurrency bug
On nvram_read, the driver searches for the variable and returns
to the user-level the address (in the mapped buffer) of the value string.
[...] The problem: suppose one user program does a read concurrently
with another doing a set (or getall). The read returns a pointer to
the value that is in the mapped buffer. Then the driver returns
control to the user level, which can cause a context switch. The other
program does something that causes a garbage collection. The read program
then de-refererences the pointer by doing a strcpy. But the pointer is now
garbage, since the GC reshuffled the buffer.

Alas, I made two attempts at the "proper" fix, and the code worked ok, but
there must be some hidden dependencies (perhaps somewhere in Broadcom code
that we don't have the source for?), because the system wouldn't come up. ;-(
This one resolves the problem but doesn't fix it right.