Add GUI for manual IPv6RD configuration - Basic/untested
Based on Stephane's 6RD proof of concept - this adds a GUI
which was 'promised' but never arrived.  Sadly 4 more
NVRAM vars (which could be shorter) Very basic/little input
validation done, but as far as I can tell it does config
a tunnel.  NVRAM defaults are set for Charter ISP