ethstate script update
when we use 3G modem in NDIS mode, then we have to change wan_ifname to new interface (ex usb0).
If WAN is not set as vlan, then ethstate broke data and etherstate will not works correct.

This commit add a simple way to detect WAN port using MAC addr.