New router support
This is a part of commit 7c33382d806ed486a19268992849f493852182cc with few updates
Big thanks for BWQ518 for this great job

K26RT-N build:
 - Tenda N6
 - Tenda N60

Builds WAS NOT TESTED! PLEASE MAKE SURE YOU ARE ABLE TO UNBRICK YOUR ROUTER if something will goes wrong.
If you not sure, please wait for official confirmation!

Started by not finished support for:
 - Buffalo WLI-H4-D1300
 - Buffalo WZR-D1100H
 - Buffalo WZR-D1800H
 - Linksys EA6500 V1

Those builds was disabled for now in Makefile.

This commit add also lot of fixes in other routers