ui: keep utilization tabs on screen
The Bandwidth and IP Traffic pages use tabs to list their interfaces and
hostnames respectively. They tabs will run off the right edge of the
window if the window is too narrow to display them all.

By floating .tab-nav, we can have it overflow and run down the page
rather than off of it.